+++ Source/EventFlow.Tests/UnitTests/EventStores/EventDefinitionServiceTests.cs
@@ -35,8 +35,11 @@ namespace EventFlow.Tests.UnitTests.EventStores
     {
         [EventVersion("Fancy", 42)]
         public class TestEventWithLongName : AggregateEvent<IAggregateRoot<IIdentity>, IIdentity> { }
+
         public class TestEvent : AggregateEvent<IAggregateRoot<IIdentity>, IIdentity> { }
+
         public class TestEventV2 : AggregateEvent<IAggregateRoot<IIdentity>, IIdentity> { }
+
         public class OldTestEventV5 : AggregateEvent<IAggregateRoot<IIdentity>, IIdentity> { }
 
         [TestCase(typeof(TestEvent), 1, "TestEvent")]
@@ -76,6 +79,13 @@ namespace EventFlow.Tests.UnitTests.EventStores
             eventDefinition.Name.Should().Be(eventName);
             eventDefinition.Version.Should().Be(eventVersion);
             eventDefinition.Type.Should().Be(expectedEventType);
+        }
+
+        [Test]
+        public void CanLoadNull()
+        {
+            // Act
+            Sut.LoadEvents(null);
         }
     }
 }