+++ RELEASE_NOTES.md
@@ -1,4 +1,18 @@
-### New in 0.13 (not released yet)
+### New in 0.14 (not released yet)
+
+* New: Created the OWIN `CommandPublishMiddleware` middleware that can
+  handle publishing of commands by posting a JSON serialized command to
+  e.g. `/commands/ping/1` in which `ping` is the command name and `1` its
+  version. Remember to add authentication
+* New: Created a new interface `ICommand<TAggregate,TIdentity,TSourceIdentity>`
+  to allow developers to control the type of `ICommand.SourceId`. Using the
+  `ICommand<TAggregate,TIdentity>` (or Command<TAggregate,TIdentity>)
+  will still yield the same result as before, i.e., `ICommand.SourceId` being
+  of type `ISourceId`
+* New: The `AddDefaults(...)` now also adds the command type definition to the
+  new `ICommandDefinitonService`
+
+### New in 0.13.962 (released 2015-09-13)
 
  * Breaking: `EventFlowOptions.AddDefaults(...)` now also adds query handlers
  * New: Added an optional `Predicate<Type>` to the following option extension
@@ -6,8 +20,12 @@
    `AddCommandHandlers(...)`, `AddDefaults(...)`, `AddEventUpgraders(...)`,
    `AddEvents(...)`, `AddMetadataProviders(...)`, `AddQueryHandlers(...)` and
    `AddSubscribers(...)`
- * Fixed: `EventFlowOptions.AddAggregateRoots(...)` now prevents abstract classes
-   from being registered when passing `IEnumerable<Type>`.
+ * Fixed: `EventFlowOptions.AddAggregateRoots(...)` now prevents abstract
+   classes from being registered when passing `IEnumerable<Type>`
+ * Fixed: Events published to RabbitMQ are now in the right order for chains
+   of subscribers, if `event A -> subscriber -> command -> aggregate -> event B`,
+   then the order of published events to RabbitMQ was `event B` and then
+   `event A`
 
 ### New in 0.12.891 (released 2015-09-04)